requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.3.0
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.21.1
cmake==4.1.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.0
django-cors-headers==4.3.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
dotenv==0.9.9
executing==2.2.0
fastapi==0.118.0
ffmpeg==1.4
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.42.0
google-genai==1.46.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.6
language_data==1.3.0
lazy_loader==0.4
loguru==0.7.3
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
openai==2.1.0
opencv-python==4.12.0.88
optuna==4.4.0
orjson==3.11.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
preshed==3.0.10
progressbar==2.5
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyglet==2.1.8
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.2
PyOpenGL==3.1.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==310
PyYAML==6.0.3
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
trimesh==4.7.4
twilio==8.13.0
typer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.37.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
win32_setctime==1.2.0
wordcloud==1.9.4
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
mediapipe
opencv-python
numpy